Minor bugfixes
Fixed a bug where certain models would crash the application in AR Mode
Now the ARViewController removes the AnchorEntity from the scene.
Removed the LatinNameLabel
Now the MainViewController shows 'Under development' for the model systems that are not done yet.
Fixed the bug when the ARViewController would crash if the model didnt have a informationText
Updated the ARViewController buttons to be responsive.